protocol: switch to PacketInfo, add more packets
Switches all of Packet subclasses for PacketInfo module
globals, as it was easier to set up with a parser.
Vector3D is now Position, as it is more accurate, and
I added type notations to a lot of the methods.

I also finished all of the required packets for getting
a basic server running, and did some basic but not very
functional testing on __main__.py

Signed-off-by: Jacob Shtabnoy <shtabnoyjacob@scps.net>