plugins: add operator checks to on_command, decorator fix
Adds a new keyword argument to on_command: 'op_only',
which when specified will only run the command if the
player is an operator.

Fixes an issue with on_start, on_shutdown,
on_player_added, and on_player_removed where the callback
was not correctly returned to python, leading to it
returning None.

Signed-off-by: Jacob Shtabnoy <shtabnoyjacob@scps.net>