server: change execution order for SERVER/NEW_PLAYER
Moves the callback 'SERVER/NEW_PLAYER' so that it will
be called after the Player is spawned into the world.
Fixes an issue where a ghost player will be spawned in
even if the player was dropped during the
'SERVER/NEW_PLAYER' callback.

Also adds a log for when new connections are opened.

Signed-off-by: Jacob Shtabnoy <shtabnoyjacob@scps.net>