plugin: add autocracy, a simple and broken admin plugin
Autocracy is a plugin which allows you to manage operators
on your server, and prevent certain players from joining
the server.

When this plugin is added to a server without any
other configuration, it will assume every player connecting
from the loopback address is an operator. In order to add
operators to the server, you must set a list on the server
object itself called _Autocracy_operators

Signed-off-by: Jacob Shtabnoy <shtabnoyjacob@scps.net>