ClassicServer: decrease compression level of maps
Decreases the compression of the level when it is sent to
newly connected players. Before it was attempting to
compress at the maximum level, and with larger files this
took several minutes.

The level has been changed so that it will only take a few
seconds to compress the level and begin sending it.

Signed-off-by: Jacob Shtabnoy <shtabnoyjacob@scps.net>